PiperOrigin-RevId: 470767654

+++ tensorflow/compiler/xla/mlir_hlo/lib/Dialect/mhlo/transforms/legalize_mhlo_to_thlo.cc
@@ -69,7 +69,7 @@ struct ConcatenateOpPattern : public OpRewritePattern<mhlo::ConcatenateOp> {
 
       // For all dimensions other than the concatenation dimension, we can copy
       // the size from any operand.
-      if (i != concatDim) {
+      if (i != static_cast<int64_t>(concatDim)) {
         dynamicInitSizes.push_back(
             rewriter.create<tensor::DimOp>(loc, anyOperand, i));
         continue;
